enabling activity pages for open groups
Enabled members to be a None type and added 'admins'. If members
is None it indicates that the group has no concept of members and
is an "open" group. Admins looks just like members but is simply
a list of the creator-this will probably be expanded at some point
to be an actual list of admins but until then it's a list of one.

Commit includes the following requirements:
- # of public annotations made in that group
- Name of the group admins (so that H users know who created and controls that group)
- Share link for group (so that H users can share the group)
- Open Groups won't appear in the groups dropdown (for now)
- H users can search for a group in the search box, and it will return any corresponding groups (using existing search rules - future changes to open group discoverability in search will be handled in another ticket).
- Open groups should have their own groups landing page (like private groups currently do - i.e. https://hypothes.is/groups/8M8734qW/artis-group)
- Open groups pages are visible to logged out users.